// src/app/@core/interfaces/auction/offices.model.ts
export interface OfficesModel {
    id?: number;
    name?: string;
    street?: string;
    noExt?: string;
    noInt?: string;
    colony?: string;
    municipalDelegate?: string;
    postalCode?: number;
    rfc?: string;
    phone?: string;
    phoneTwo?: string;
    fax?: string;
    typeOffice?: string;
    noRegistration?:number;
}